fix ws connections

+++ backend/src/controller/WebsocketController.ts
@@ -1,34 +1,34 @@
 import { Request } from "express";
 import { WebSocket } from "ws";
 import { database, sessionService, socketService } from "..";
-import { UserWithSession } from "../models/SessionModels";
+import { exportUser } from "../models/SessionModels";
 import { SocketMessage } from "../models/WebsocketModels";
 import { SessionController } from "./SessionController";
 import { UserController } from "./UserController";
 
-async function authenticateSocket(socket, req): Promise<UserWithSession> {
+async function authenticateSocket(socket, req): Promise<number> {
 	if (!req.headers["user-id"]) {
 		socket.close();
 		return;
 	}
 
-	const user = await database.user.findFirst({
+	const user = await database.user.findUnique({
 		where: {
-			internal_id: req.headers["user-id"].tostring()
+			spotify_id: req.headers["user-id"]
 		},
 		include: { session: true }
 	});
 	if (!user) {
-		socket.close();
+		socket.close(1008, `User ${req.headers["user-id"]} does not exist`);
 		return;
 	}
 	await socketService.addConnection(user.id, socket);
-	return user;
+	return user.id;
 }
 
 export async function handleConnection(ws: WebSocket, req: Request) {
 	// Authenticate and add to our persistant set of connections
-	const currentUser = await authenticateSocket(ws, req);
+	const currentUserId = await authenticateSocket(ws, req);
 
 	ws.on('error', console.error);
 
@@ -40,10 +40,10 @@ export async function handleConnection(ws: WebSocket, req: Request) {
 			} else {
 				switch (req.method) {
 					case "SESSION":
-						new SessionController().acceptRequest(ws, req, currentUser);
+						new SessionController().acceptRequest(ws, req, currentUserId);
 						break;
 					case "PLAYING":
-						new UserController().updateCurrentlyPlaying(ws, req, currentUser);
+						new UserController().updateCurrentlyPlaying(ws, req, currentUserId);
 						break;
 					default:
 						ws.send(JSON.stringify({ Error: "Received data has an invalid method"}));
@@ -57,8 +57,13 @@ export async function handleConnection(ws: WebSocket, req: Request) {
 	ws.on('close', async function close(code, reason) {
 		console.error(`Socket Closed: ${reason.toString()}`);
 		const userId = await socketService.retrieveBySocket(ws);
-		await sessionService.leaveSession(userId);
-		await socketService.removeConnectionBySocket(ws);
+		if (userId) {
+			const session = await sessionService.leaveSession(userId);
+			if (session) {
+				await sessionService.messageSession(session.id, currentUserId, { userLeave: exportUser(session.members.find(x => x.id === currentUserId)) });
+			}
+			await socketService.removeConnectionBySocket(ws);
+		}
 	});
 }
